Move ErrorCode to twenty-shared
