+++ packages/twenty-server/test/integration/graphql/suites/settings-permissions/data-model.integration-spec.ts
@@ -10,10 +10,9 @@ import { deleteOneObjectMetadataItemFactory } from 'test/integration/metadata/su
 import { deleteOneObjectMetadataItem } from 'test/integration/metadata/suites/object-metadata/utils/delete-one-object-metadata.util';
 import { updateOneObjectMetadataItemFactory } from 'test/integration/metadata/suites/object-metadata/utils/update-one-object-metadata-factory.util';
 import { makeMetadataAPIRequestWithMemberRole } from 'test/integration/metadata/suites/utils/make-metadata-api-request-with-member-role.util';
-import { FieldMetadataType } from 'twenty-shared';
+import { ErrorCode, FieldMetadataType } from 'twenty-shared';
 
 import { SEED_APPLE_WORKSPACE_ID } from 'src/database/typeorm-seeds/core/workspaces';
-import { ErrorCode } from 'src/engine/core-modules/graphql/utils/graphql-errors.util';
 import { PermissionsExceptionMessage } from 'src/engine/metadata-modules/permissions/permissions.exception';
 
 describe('datamodel permissions', () => {
