+++ packages/twenty-server/test/integration/graphql/suites/settings-permissions/workspace-invitation.integration-spec.ts
@@ -1,9 +1,9 @@
 import request from 'supertest';
 import { makeGraphqlAPIRequest } from 'test/integration/graphql/utils/make-graphql-api-request.util';
 import { updateFeatureFlagFactory } from 'test/integration/graphql/utils/update-feature-flag-factory.util';
+import { ErrorCode } from 'twenty-shared';
 
 import { SEED_APPLE_WORKSPACE_ID } from 'src/database/typeorm-seeds/core/workspaces';
-import { ErrorCode } from 'src/engine/core-modules/graphql/utils/graphql-errors.util';
 import { PermissionsExceptionMessage } from 'src/engine/metadata-modules/permissions/permissions.exception';
 
 const client = request(`http://localhost:${APP_PORT}`);
@@ -112,7 +112,7 @@ describe('workspace invitation permissions', () => {
     const queryData = {
       query: `
         mutation deleteWorkspaceInvitation {
-          deleteWorkspaceInvitation(appTokenId: "test-invitation-id") 
+          deleteWorkspaceInvitation(appTokenId: "test-invitation-id")
         }
       `,
     };
